requirements.txt
opencv-python~=4.9.0.80
pillow~=10.3.0
numpy~=1.26.4
Cython~=3.0.10
setuptools~=68.2.0
